react router and pagination to posts

+++ src/forum-js/components/forum-form/ForumPosts.js
@@ -1,9 +1,23 @@
 import React from 'react';
 import { connect } from 'react-redux';
+import { Link } from 'react-router';
 
 import * as asyncActions from '../../actions/async';
 
 class ForumPosts extends React.Component {
+
+	componentDidMount(){
+		const { props } = this;
+		props.getPosts(props.pageNumber);
+	}
+
+	componentWillReceiveProps(nextProps){
+		const { props } = this;
+		if (props.pageNumber !== nextProps.pageNumber){
+			props.getPosts(nextProps.pageNumber);
+		}
+	}
+
 
 	_deletePost(postId){
 		console.log(postId);
@@ -20,12 +34,18 @@ class ForumPosts extends React.Component {
 	render(){
 		const { props } = this;
 
+		console.log('props.postsprops.postsprops.posts', props.postsTotalCount);
+
+		const pagesCount = Math.ceil(props.postsTotalCount / 5);
+
+		console.log(pagesCount);
+
 		return(
 			<div className={( (props.mixClass ? props.mixClass : '') + ' posts')}>
 
 				<ul className="posts__list">
 
-					{props.posts && props.posts.map( (post,i) => {
+					{props.posts.map( (post,i) => {
 
 						const value = JSON.parse(decodeURIComponent(post.Value));
 
@@ -33,18 +53,18 @@ class ForumPosts extends React.Component {
 
 						let deleteIt;
 
-						//if (props.profile.roles.indexOf('EduStaff') > -1){
+						if (props.profile.roles.indexOf('EduStaff') > -1){
 							deleteIt = (
 								<button
 									onClick={this._deletePostHandler(post.Key)}
 								>Удалить</button>
 							);
-						//} 
+						}
 
-						console.log(post, value);
+						//console.log(post, value);
 
 						return (
-						
+
 							<li className={('posts__item post ' + (value.user.id === '1000001035607' ? 'post--super' : '') )} key={i}>
 
 								<div className="post__avatar-placeholder">
@@ -96,11 +116,44 @@ class ForumPosts extends React.Component {
 								{deleteIt}
 
 							</li>
-						
+
 						)
 					})}
 
 				</ul>
+
+				<div className="posts__pagination pagination">
+
+					<ul className="pagination__list">
+
+						{[,...Array(pagesCount)].map( (value, i)  => {
+
+							let link = '/page/' + (i);
+
+							if (i === 1){
+								link = '';
+							}
+
+							return (
+
+							<li className="pagination__item" key={i}>
+
+								<Link 
+									to={link}
+									activeClassName="active"
+									onlyActiveOnIndex={(i === 1)}
+								>
+									{i}
+								</Link>
+
+							</li>							
+
+							);
+						})}
+						
+					</ul>
+
+				</div>
 
 			</div>
 		);
@@ -110,11 +163,13 @@ class ForumPosts extends React.Component {
 
 const mapStateToProps = (state, ownProps) => ({
 	profile: state.user.profile,
-	posts: state.posts.Keys,
+	posts: state.posts.list,
+	postsTotalCount: state.posts.itemsTotalCount,
 });
 
 const mapDispatchToProps = (dispatch, ownProps) => ({
 	deletePost: (postId) => dispatch(asyncActions.deletePost(postId)),
+	getPosts: (pageNumber) => dispatch(asyncActions.getPosts(pageNumber)),
 });
 
 ForumPosts.propTypes = {
